requirements.txt
attrs==23.2.0
blinker==1.8.2
Bottleneck==1.4.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
filelock==3.15.4
firebase-admin==6.5.0
flasgger==0.9.7b2
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.8.1.78
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pynput==1.7.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
rsa==4.9
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sounddevice==0.4.7
sympy==1.13.0
tbb==2021.13.0
torch==2.3.1+cpu
torchaudio==2.3.1
torchvision==0.18.1+cpu
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.0.163
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
